requirements.txt

altgraph==0.17
appdirs==1.4.4
bitstring==3.1.7
blabel==0.1.6
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cmsis-pack-manager==0.3.0
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.16.1
future==0.18.2
hidapi==0.10.1
idna==2.10
intelhex==2.3.0
intervaltree==3.1.0
iso8601==1.0.2
joblib==1.0.1
milksnake==0.1.5
naturalsort==1.5.1
nltk==3.6.2
numpy==1.21.2
pandas==1.3.3
pefile==2019.4.18
prettytable==2.2.0
psutil==5.8.0
pycodestyle==2.7.0
pycparser==2.20
pyelftools==0.27
pyinstaller-hooks-contrib==2021.1
pyinstaller==4.3
pylink-square==0.10.1
pyocd-pemicro==1.0.6
pypemicro==0.1.7
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.1
pyusb==1.2.1
pywin32-ctypes==0.2.0
pywsd==1.2.4
PyYAML==5.4.1
reedsolo==1.5.4
regex==2021.8.28
requests==2.25.1
six==1.15.0
sortedcontainers==2.4.0
toml==0.10.2
tqdm==4.62.2
urllib3==1.26.4
wcwidth==0.2.5
wn==0.8.0